// types.ts
export interface IPFSAsset {
  id: string;
  name: string;
  type: 'Image' | 'Audio' | 'Video' | 'Document';
  size: string;
  cid: string;
  url: string;
  status: 'Uploading' | 'Pinned' | 'Minted';
  timestamp: string;
}